treewide: Align networking code with net_mgmt API change
The net_mgmt event representation changed from 32-bit to 64-bit value,
therefore update all net_mgmt callbacks to align with this change.

+++ samples/cellular/nrf_cloud_rest_cell_location/src/main.c
@@ -548,7 +548,7 @@ int init(void)
 
 /* Callback to track network connectivity */
 static struct net_mgmt_event_callback l4_callback;
-static void l4_event_handler(struct net_mgmt_event_callback *cb, uint32_t event,
+static void l4_event_handler(struct net_mgmt_event_callback *cb, uint64_t event,
 			     struct net_if *iface)
 {
 	if ((event & CONN_LAYER_EVENT_MASK) != event) {
@@ -570,7 +570,7 @@ static void l4_event_handler(struct net_mgmt_event_callback *cb, uint32_t event,
 /* Callback to track connectivity layer events */
 static struct net_mgmt_event_callback conn_cb;
 static void connectivity_event_handler(struct net_mgmt_event_callback *cb,
-					   uint32_t event,
+					   uint64_t event,
 					   struct net_if *iface)
 {
 	if (event == NET_EVENT_CONN_IF_FATAL_ERROR) {
