+++ samples/cellular/nrf_provisioning/src/main.c
@@ -136,7 +136,7 @@ static struct nrf_provisioning_dm_change dmode = { .cb = device_mode_cb, .user_d
 /* Callback to track network connectivity */
 static struct net_mgmt_event_callback l4_callback;
 static void l4_event_handler(struct net_mgmt_event_callback *cb,
-			     uint32_t event, struct net_if *iface)
+			     uint64_t event, struct net_if *iface)
 {
 	if ((event & EVENT_MASK) != event) {
 		return;
